+++ tests/test_domains/test_domain_py.py
@@ -480,6 +480,13 @@ def test_no_index_entry(app):
     assert_node(doctree[0], addnodes.index, entries=[('single', 'f (built-in class)', 'f', '', None)])
     assert_node(doctree[2], addnodes.index, entries=[])
 
+    text = (".. py:module:: h\n"
+            ".. py:module:: i\n"
+            "   :no-index-entry:\n")
+    doctree = restructuredtext.parse(app, text)
+    assert_node(doctree, (addnodes.index, nodes.target, nodes.target))
+    assert_node(doctree[0], addnodes.index, entries=[('pair', 'module; h', 'module-h', '', None)])
+
 
 @pytest.mark.sphinx('html', testroot='domain-py-python_use_unqualified_type_names')
 def test_python_python_use_unqualified_type_names(app, status, warning):